refactor: logic to open external links in new tab (#173)
* fix: open external links in new tab

Fix applies to BannerHeader and NavPrimary. Logic is consolidated and refactored in SmartLink. The explicit target from the backend will take precedence over URL analysis in NavPrimary, but not (yet?) in BannerHeader, as we don't currently have a 'target' in our graphql query.

* refactor: move new-tab logic to SmartLink

* fix: BlockHighligh link targets

was always opening links in new tab because the computed value `isImpactReport()` returned `"false"` (which is truthy) instead of `false`

* fix: disable domain check for isInternalLink

It will need to be made customizable for microsites

* fix! rename `target` props to `linkTarget`

Because the prop name `target` conflicts with the html property of the same name, the default prop value of "" from parent components was overwriting SmartLink's computed value of "_blank" for external links.

* fix: one more `target` -> `linkTarget`

Co-authored-by: Parinita Mulak <pghorpade@library.ucla.edu>