feat: ModalLightbox component (#243)
Tried to adapt something that could be used for both MediaGallery and Campus Map layouts, but that proved too difficult. This is the Map layout, and can be generalized to other cases. Main content is inserted in a slot, and presented in the middle 75% of each window axis.

Also fixes an error warning about loading SectionHeader in FlexibleBlocks